# backend/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MedFinder - Medicine Tracker API",
    "version": "1.0.0",
    "description": "Full REST API for MedFinder — Local Medicine Availability Tracker. Supports patient, pharmacy, and admin roles."
  },
  "servers": [{ "url": "http://localhost:5000/api", "description": "Development server" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/medicines": {
      "post": {
        "tags": ["Medicines"],
        "summary": "Create a new medicine (Admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "category"],
                "properties": {
                  "name": { "type": "string", "example": "Paracetamol" },
                  "genericName": { "type": "string" },
                  "category": { "type": "string", "example": "Analgesic" },
                  "form": { "type": "string", "example": "Tablet" },
                  "strength": { "type": "string", "example": "500mg" },
                  "manufacturer": { "type": "string" },
                  "requiresPrescription": { "type": "boolean", "default": false },
                  "mrp": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Medicine created" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/pharmacies": {
      "get": {
        "tags": ["Pharmacies"],
        "summary": "Get all approved pharmacies",
        "security": [],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "city", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of pharmacies" }
        }
      }
    },
...
    "/stock": {
      "get": {
        "tags": ["Stock"],
        "summary": "Get own pharmacy's stock list (Pharmacy/Admin)",
        "responses": {
          "200": { "description": "Stock items" }
        }
      },
      "post": {
        "tags": ["Stock"],
        "summary": "Add a new stock item (Pharmacy/Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["medicine", "quantity", "price", "expiryDate"],
                "properties": {
                  "medicine": { "type": "string", "description": "Medicine ID" },
                  "quantity": { "type": "integer", "example": 100 },
                  "price": { "type": "number", "example": 45.5 },
                  "discount": { "type": "number", "example": 5 },
                  "expiryDate": { "type": "string", "format": "date", "example": "2026-12-31" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Stock item added" }
        }
      }
    },
...
    "/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create a new order (Patient only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["pharmacyId", "items", "deliveryType"],
                "properties": {
                  "pharmacyId": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "medicineId": { "type": "string" },
                        "quantity": { "type": "integer" }
                      }
                    }
                  },
                  "deliveryType": { "type": "string", "enum": ["delivery", "pickup"] },
                  "paymentMethod": { "type": "string", "enum": ["online", "cod"] },
                  "deliveryAddress": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created" }
        }
      }
    },
...
    "/alerts": {
      "get": {
        "tags": ["Alerts"],
        "summary": "Get all alerts for the authenticated user",
        "responses": {
          "200": { "description": "List of alerts" }
        }
      },
      "post": {
        "tags": ["Alerts"],
        "summary": "Create a new stock alert for a medicine",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["medicineId"],
                "properties": {
                  "medicineId": { "type": "string" },
                  "pharmacyId": { "type": "string", "description": "Optional — alert for specific pharmacy" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Alert created" }
        }
      }
    },
...
    "/reviews": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get all reviews (public)",
        "security": [],
        "responses": {
          "200": { "description": "All reviews" }
        }
      },
      "post": {
        "tags": ["Reviews"],
        "summary": "Create a new review",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["pharmacyId", "rating"],
                "properties": {
                  "pharmacyId": { "type": "string" },
                  "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Review created" }
        }
      }
    },
...
    "/saved-medicines": {
      "get": {
        "tags": ["Saved Medicines"],
        "summary": "Get all saved/bookmarked medicines (Patient only)",
        "responses": {
          "200": { "description": "Saved medicines list" }
        }
      },
      "post": {
        "tags": ["Saved Medicines"],
        "summary": "Save/bookmark a medicine (Patient only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["medicineId"],
                "properties": {
                  "medicineId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Medicine saved" }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Users (Admin)"],
        "summary": "Get all users (Admin only)",
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["patient", "pharmacy", "admin"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "List of users" }
        }
      }
    },
...
  }
}